Refactor AWS configuration and infrastructure
