+++ video-webapp/docker-compose.yml
@@ -6,18 +6,7 @@ services:
       - "4000"
     environment:
       - AWS_REGION=${AWS_REGION:-ap-southeast-2}
-      - CLIENT_ORIGINS=${CLIENT_ORIGINS:-http://localhost:5173}
-      - S3_BUCKET=${S3_BUCKET:-local-video-bucket}
-      - DYNAMO_TABLE=${DYNAMO_TABLE:-LocalVideos}
-      - DYNAMO_OWNER_INDEX=${DYNAMO_OWNER_INDEX:-OwnerIndex}
-      - COGNITO_USER_POOL_ID=${COGNITO_USER_POOL_ID:-ap-southeast-2_CdVnmKfrW}
-      - COGNITO_CLIENT_ID=${COGNITO_CLIENT_ID:-11pap5u5svkhr1hgjf934sj0id}
-      - S3_RAW_PREFIX=${S3_RAW_PREFIX:-raw-videos/}
-      - S3_TRANSCODED_PREFIX=${S3_TRANSCODED_PREFIX:-transcoded-videos/}
-      - S3_THUMBNAIL_PREFIX=${S3_THUMBNAIL_PREFIX:-thumbnails/}
-      - PRESIGNED_TTL_SECONDS=${PRESIGNED_TTL_SECONDS:-900}
-      - LIMIT_FILE_SIZE_MB=${LIMIT_FILE_SIZE_MB:-512}
-      - JWT_SECRET=${JWT_SECRET:-video-webapp-dev-secret}
+      - PORT=${PORT:-4000}
     networks:
       - appnet
 
@@ -26,10 +15,10 @@ services:
     expose:
       - "80"
     environment:
-      - VITE_API_URL=${VITE_API_URL:-/api}
-      - VITE_AWS_REGION=${VITE_AWS_REGION:-ap-southeast-2}
-      - VITE_COGNITO_USER_POOL_ID=${VITE_COGNITO_USER_POOL_ID:-ap-southeast-2_CdVnmKfrW}
-      - VITE_COGNITO_CLIENT_ID=${VITE_COGNITO_CLIENT_ID:-11pap5u5svkhr1hgjf934sj0id}
+      - VITE_API_URL=${VITE_API_URL}
+      - VITE_AWS_REGION=${AWS_REGION:-ap-southeast-2}
+      - VITE_COGNITO_USER_POOL_ID=${VITE_COGNITO_USER_POOL_ID}
+      - VITE_COGNITO_CLIENT_ID=${VITE_COGNITO_CLIENT_ID}
     networks:
       - appnet
